Move timestampStatsOfIndex() into EsIndexRangeService to resolve circular dependency
