+++ graylog2-server/src/test/resources/org/graylog2/indexer/ranges/EsIndexRangeServiceTest-EmptyIndex.json
